backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
python-dotenv>=1.0.0
asyncpg>=0.29.0
PyJWT>=2.8.0
slowapi>=0.1.9
